# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stock api",
        "description": "REST API приложения для справочника Организаций, Зданий, Деятельности",
        "contact": {
            "email": "vahagn99ghukasyan@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities/{id}/organizations": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить список организаций по ID вида деятельности",
                "description": "Возвращает список организаций, связанных с указанной деятельностью",
                "operationId": "getOrganizationsByActivity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API key required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить список всех зданий",
                "description": "Возвращает список зданий из справочника",
                "operationId": "getBuildings",
                "responses": {
                    "200": {
                        "description": "Список зданий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Building"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API key required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/buildings/{id}/organizations": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить организации по ID здания",
                "description": "Возвращает список организаций, находящихся в указанном здании",
                "operationId": "getOrganizationsByBuilding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API key required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организацию по ID",
                "description": "Возвращает организацию по ID",
                "operationId": "getOrganizationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API key required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/organizations/search": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций",
                "description": "Поиск организаций по тексту, координатам, адресу здания и активности.",
                "operationId": "getOrganizationsBySearch",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Поисковый текст (например, часть названия)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Поликлиника"
                        }
                    },
                    {
                        "name": "coordinates[latitude]",
                        "in": "query",
                        "description": "Широта для геопоиска",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 40.1792
                        }
                    },
                    {
                        "name": "coordinates[longitude]",
                        "in": "query",
                        "description": "Долгота для геопоиска",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 44.4991
                        }
                    },
                    {
                        "name": "coordinates[radius]",
                        "in": "query",
                        "description": "Радиус поиска в метрах",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1000
                        }
                    },
                    {
                        "name": "building[address]",
                        "in": "query",
                        "description": "Адрес здания",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ул. Пушкина, д. 10"
                        }
                    },
                    {
                        "name": "activity[name]",
                        "in": "query",
                        "description": "Название вида деятельности",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Медицина"
                        }
                    },
                    {
                        "name": "activity[deep_search]",
                        "in": "query",
                        "description": "Глубокий поиск по видам деятельности",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API key required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Молочная продукция"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "Building": {
                "title": "Building",
                "required": [
                    "id",
                    "address",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина 1"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6173
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Organization",
                "required": [
                    "id",
                    "name",
                    "building_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Рога и Копыта"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "8-900-123-45-67"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Авторизация по заголовку: Authorization: Api-Key {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Activities",
            "description": "Activities"
        },
        {
            "name": "Buildings",
            "description": "Buildings"
        },
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}